Прочитав код, фикшу вещи, что могут стать ошибками

+++ bot.js
@@ -90,16 +90,19 @@ client.on('message', message => {
         };
     };
 
-    if (command[0] == `${prefix}идти` && mb == false && message.channel.name == 'улица'){
+    if (command[0] == `${prefix}идти` && mb == false){
         message.delete();
 
-        if (command[1] == "на"){
+        if (command[1] == "на" && message.channel.name == 'улица'){
             let homestreet = street.find(st => st.name == message.channel.parent.name);
             let walkway = homestreet.radius.find(st => st.toLowerCase() == args.toLowerCase());
 
-            if (walkway != null && message.channel.permissionOverwrites.get(message.author.id) != null){
-                client.channels.cache.find(cat => cat.name == walkway).updateOverwrite(message.author, { VIEW_CHANNEL: true });
-                message.channel.parent.permissionOverwrites.get(message.author.id).delete();
+            if (walkway != null && message.channel.parent.permissionOverwrites.get(message.author.id) != null){
+                let cat = client.channels.cache.find(cat => cat.name == walkway);
+                if (cat.type == 'category'){
+                    client.channels.cache.find(cat => cat.name == walkway).updateOverwrite(message.author, { VIEW_CHANNEL: true });
+                    message.channel.parent.permissionOverwrites.get(message.author.id).delete();
+                };
             }else if (walkway == null && street.find(st => st.name.toLowerCase() == args.toLowerCase()) != null){
                 message.author.send(`${args} не является соседней улицей с ${homestreet.name}.`);
             }else{
